v1.0-rc.1 iteration: invariants + JOB-Slow real + ablation runner + scaffolds
Source-side closure of the v1.0 iteration items surfaced by Wave 3/4
empirical hardening. Headline:

- SRC01: try_new added to all 5 sketches (HLL/Bloom/CMS/EquiDepth/Correlated2D)
  with parameter-validation-before-allocation. validate() invariant checks in
  every from_bytes (H04 16-byte all-zero HLL leak + EquiDepth zero-payload
  leak now rejected). BloomFilter::new clamps instead of allocating multi-EiB
  on fp_rate=0 (H02). 29 new unit tests.

- SRC03: 47 doctests across constructors, LpBound, Puffin round-trip,
  corrector trait, ColumnStats, FeedbackStore (0 -> 52.2% coverage).

- WAVE4-B: parse test for all 113 JOB-Slow queries verified byte-identical
  to Leis VLDB 2015 canonical corpus (winkyao/join-order-benchmark
  @ 8e337db).

- WAVE4-C/F: JOB-Slow real end-to-end now executes against IMDb. Tokio
  nested-runtime panic fixed via register_imdb_tables_async.
  SamkhyaTableProvider wired into the IMDb path with sidecar-aware
  fallback. build-puffin extended with --imdb-dir to produce HLL+Bloom
  sidecars for the 21 IMDb tables.

- WAVE4-D: fetch_imdb.sh default URL flipped to live CWI mirror with
  SHA-256 fail-fast. SamkhyaTableProvider overrides for TPC-H tables.
  calibrate --tpch-dir widening + tpch_smoke.rs integration test.

- WAVE4-E: L4 GbtCorrector feature-expanded (5 features: baseline_estimate
  + min_table_cardinality + join_key_skew_ratio + chainbound_ceiling_log10
  + prior_residual_log) with 6-pass warmup. Partial recovery: regression
  dropped +386% -> +137% but still BH-significant; L4/L5 deferred to v1.1.
  Production deployment recommendation: A2 (L1+L2+L3) only.

- New ablation_runner binary (5x10x30 grid, BCa CIs, paired Wilcoxon, BH FDR).

- Statistical scripts: bootstrap_ci.py --method bca (Efron-Tibshirani 1993
  ch. 14, n=10000), wilcoxon_paired.py (Wilcoxon 1945), benjamini_hochberg.py
  (BH 1995). Pure-stdlib, exit-zero smoke tests.

- Fortress tests added for arrow/duckdb/iceberg/polars; 5 new cargo-fuzz
  targets for sketch decoders.

- New samkhya-it integration crate scaffold. samkhya-py: pyi typing
  stubs + test_sketches.py.

- samkhya-duckdb-ext: cxx wrapper.cc/wrapper.h replaces deleted
  extension.cpp (staticlib+rlib only per v1.0 design;
  cdylib + runtime LOAD waits on DuckDB Issue #11638).

Workspace gates green: cargo check, clippy -D warnings, fmt --check, test
(0 failed across all reported test results). samkhya-postgres excluded
from workspace clippy/test only because pgrx 0.12 bindgen panics on
simultaneous pg13..pg17 features (pre-existing).

Strategic docs (README/CHANGELOG/SECURITY/PR template + per-crate READMEs +
REPRODUCIBILITY.md) held for separate per-commit approval.
v1.0.0 release tag held pending explicit user confirmation.

+++ samkhya-core/examples/sketch_to_puffin.rs
@@ -81,7 +81,7 @@ fn main() -> Result<()> {
     for &probe in &[0u64, 1, 42, 4_999, 5_000, 9_999, 1_000_000] {
         let bytes = probe.to_le_bytes();
         let present = bloom_back.contains(&bytes);
-        let truly_present = (probe as u64) < distinct_target;
+        let truly_present = probe < distinct_target;
         println!("  id={probe:>8}  bloom_says={present:<5}  actually_inserted={truly_present}");
     }
